requirements.txt
streamlit
google-generativeai
sentence-transformers
deep-translator
pdfplumber
docx2txt
torch
python-dotenv
